sprite's X and Y position are now doubles instead of ints for greater precision

+++ source/sprite.hpp
@@ -317,8 +317,8 @@ class Sprite {
         bool toDelete;
         int currentCostume;
         int volume;
-        int xPosition;
-        int yPosition;
+        double xPosition;
+        double yPosition;
         int size;
         int rotation;
         int layer;
